vue 2.x商城

+++ src/views/cart/childComps/CartListItem.vue
@@ -1,5 +1,6 @@
 <template>
-  <div class="cart-list-item" :class="{ 'touch-move-active' : product.isTouchMove }">
+  <!-- :class="{ 'touch-move-active' : product.isTouchMove }" -->
+  <div class="cart-list-item">
     <div class="content">
       <div class="item-selector">
         <CheckButton :isChecked="product.checked" @click.native="checkClick" />
@@ -12,11 +13,11 @@
         <div class="item-desc">商品描述: {{ product.desc }}</div>
         <div class="info-bottom">
           <div class="item-price">¥{{ product.price }}</div>
-          <InputNumber class="item-count" :counter="product.count" @numChange="changeCount" />
+          <InputNumber class="item-count" :counter="product.count" @numChange="changeCount" @deleteItem="deleteClick" />
         </div>
       </div>
     </div>
-    <div class="del" @click="deleteClick">删除</div>
+    <!-- <div class="del" @click="deleteClick">删除</div> -->
   </div>
 </template>
 
@@ -41,18 +42,29 @@ export default {
       type: Number,
       default: 0
     },
-		isTouchMove: { // 是否显示删除
-			type:Boolean,
-			default: false
-		}
-	},
+		// isTouchMove: { // 是否显示删除
+		// 	type:Boolean,
+		// 	default: false
+		// }
+	}, 
+  watch: {
+    "product.count": {
+      immediate: true,
+      handler(val) {
+        this.$forceUpdate();
+      }
+    }
+  },
 	methods: {
 		...mapMutations(['selectItem','deleteItem','updateProduct']),
 		checkClick() {
       this.selectItem(this.index);
     },
     deleteClick() {
-      this.deleteItem(this.index);
+      const isConfirm = window.confirm(`你确定要删除${this.product.title}吗?`);
+      if (isConfirm) {
+        this.deleteItem(this.index);
+      }
     },
     changeCount(val) {
       this.updateProduct({
@@ -73,9 +85,9 @@ export default {
     width: 100%;
     padding: 5px;
     border-bottom: 1px solid $border-ccc;
-    transition: all 0.4s;
-		transform: translateX(80px);
-		margin-left:-80px;
+    // transition: all 0.4s;
+		// transform: translateX(80px);
+		// margin-left:-80px;
 
     .item-selector {
       display: flex;
@@ -118,21 +130,21 @@ export default {
       }
     }
   }
-  .del {
-    background-color: $bg-tint;
-    width: 90px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    color: $text-white;
-    transform: translateX(80px);
-    transition: all 0.4s;
-	}
+  // .del {
+  //   background-color: $bg-tint;
+  //   width: 90px;
+  //   display: flex;
+  //   flex-direction: column;
+  //   align-items: center;
+  //   justify-content: center;
+  //   color: $text-white;
+  //   transform: translateX(80px);
+  //   transition: all 0.4s;
+	// }
 	
-		&.touch-move-active .content,
-		&.touch-move-active .del {
-		transform: translateX(0);
-	}
+	// 	&.touch-move-active .content,
+	// 	&.touch-move-active .del {
+	// 	transform: translateX(0);
+	// }
 }
 </style>
